renamed security folder into auth
