+++ passik/application/src/main/java/port/in/auth/AuthenticationUseCase.java
@@ -1,0 +1,9 @@
+package port.in.auth;
+
+import de.daycu.passik.model.auth.MasterLogin;
+import de.daycu.passik.model.auth.MasterPassword;
+import service.auth.AuthenticationResult;
+
+public interface AuthenticationUseCase {
+    AuthenticationResult authenticate(MasterLogin masterLogin, MasterPassword masterPassword);
+}
