+++ passik/model/src/main/java/de/daycu/passik/model/auth/Salt.java
@@ -1,4 +1,4 @@
-package de.daycu.passik.model.security;
+package de.daycu.passik.model.auth;
 
 import org.apache.shiro.crypto.RandomNumberGenerator;
 import org.apache.shiro.crypto.SecureRandomNumberGenerator;
